update session moved from login controller to UserDetailsServiceImpl
